TransactionList is modular component
I removed the repeated code from employee dashboard, now I call the page(renamed TransactionList) in its employee dashboard part, giving it a variable that changes the source of the transactions.

this makes the code more organized, while alsso making it easier for Tabeeb to use my page for his user story

also I believe this structure would be a lot more helpful for any employee dashboard sub-page
